feat: closes #18 and closes #21

+++ utils/fs.js
@@ -32,23 +32,61 @@ const read = (relativePath) => {
 //   }
 // }
 
-const write = async (relativePath, content, options = { format: true }) => {
+const write = async (
+  relativePath,
+  content,
+  options = { format: true, force: false },
+) => {
   try {
     const absPath = pathJoin(relativePath);
     if (options.format) {
       content = await format(content, options?.parser);
     }
-    fs.writeFileSync(absPath, content);
-    console.log(chalk.bgGreen` Update `, ' ', relativePath);
+    // Check if file exists and read its content
+    let existingContent = '';
+    if (fs.existsSync(absPath)) {
+      existingContent = fs.readFileSync(absPath, 'utf8');
+    }
+    // if (relativePath == 'models/index.js') {
+    //   console.log('NEW Content -----', content)
+    // }
+    if (content !== existingContent) {
+      fs.writeFileSync(absPath, content);
+      console.log(chalk.blue`UPDATE `, ' ', '/' + relativePath);
+    } else {
+      console.log(
+        chalk.yellow`SKIP  `,
+        ' ',
+        '/' + relativePath,
+        '(no changes)',
+      );
+    }
   } catch (error) {
     console.error('FAILED WRITING TO FILE ', relativePath, error);
   }
 };
 
-const append = (relativePath, content) => {
+const append = async (relativePath, content, options = { format: true }) => {
   try {
     const absPath = pathJoin(relativePath);
-    fs.appendFileSync(absPath, content);
+    let existingContent = '';
+    if (fs.existsSync(absPath)) {
+      existingContent = fs.readFileSync(absPath, 'utf8');
+    }
+    if (!existingContent.includes(content)) {
+      if (options.format) {
+        content = await format(content, options?.parser);
+      }
+      fs.appendFileSync(absPath, '\n' + content);
+      console.log(chalk.blue`UPDATE `, ' ', '/' + relativePath);
+    } else {
+      console.log(
+        chalk.yellow`SKIP  `,
+        ' ',
+        '/' + relativePath,
+        '(no changes)',
+      );
+    }
   } catch {
     console.error('FAILED WRITING TO FILE ', relativePath);
   }
@@ -62,7 +100,7 @@ const exists = (path) => {
 const createDirectory = (path) => {
   if (exists(path)) return;
   fs.mkdirSync(path);
-  console.log(chalk.bgGreen` Create `, ' ', path);
+  console.log(chalk.green`CREATE `, ' ', '/' + path);
 };
 
 const saveConfig = (data) => {
